exception handling and few changes

+++ mHealthSystem/src/main/java/com/cg/mHealthSystem/Controllers/AdminController.java
@@ -1,5 +1,7 @@
 package com.cg.mHealthSystem.Controllers;
 
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -8,6 +10,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.cg.mHealthSystem.Exception.ResourceNotFoundException;
+import com.cg.mHealthSystem.Repository.DepartmentRepository;
+import com.cg.mHealthSystem.Repository.DoctorRepository;
+import com.cg.mHealthSystem.Repository.NurseRepository;
+import com.cg.mHealthSystem.Repository.PatientDetailsRepository;
 import com.cg.mHealthSystem.entity.Department;
 import com.cg.mHealthSystem.entity.Doctor;
 import com.cg.mHealthSystem.entity.Nurse;
@@ -18,6 +25,16 @@ import com.cg.mHealthSystem.services.AdminService;
 public class AdminController {
 	@Autowired
 	private AdminService adminservice ;
+	
+	@Autowired
+	private  DoctorRepository  doctorDao;
+	@Autowired
+	private NurseRepository nurseDao;
+	
+	@Autowired
+	private DepartmentRepository departmentDao ;
+
+	private PatientDetailsRepository patientDao;
 	
 	@PostMapping("/addDoctor")
 	public Doctor addDoctor(@RequestBody Doctor doctor )
@@ -33,7 +50,17 @@ public class AdminController {
 	@DeleteMapping("/deleteNurse/{nurseId}")
 	public boolean deleteNurse(@PathVariable Integer nurseId)
 	{
-		return adminservice.removeNurseById(nurseId);
+		Optional<Nurse> nurse = nurseDao.findById(nurseId);
+		if(nurse.isPresent())
+		{
+			return adminservice.removeNurseById(nurseId);
+		}
+		else
+		{
+			throw new ResourceNotFoundException();
+		}
+		
+		
 	}
 	@DeleteMapping("/deleteDoctor/{doctorId}")
 	public boolean deleteDoctor(@PathVariable Integer doctorId)
@@ -50,7 +77,13 @@ public class AdminController {
 	@DeleteMapping("/deleteDepartment/{deptId}")
 	public boolean deleteDepartment(@PathVariable Integer deptId)
 	{
-		return adminservice.removeDepartment(deptId);
+		Optional<Department> department = departmentDao.findById(deptId);
+		if(department.isPresent())
+		{
+			return adminservice.removeDepartment(deptId);
+		}
+		
+		throw new ResourceNotFoundException();
 	}
 	
 	@DeleteMapping("/deletePatient/{patientId}")
